\name{critera}
\alias{critera}
\title{
Criterion of the Tucker3 algorithm
}
\description{
The function critera is used to define the criterion of the Tucker3 algorithm.
}
\usage{
critera(aold, anew)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{aold}{
The old component of the first mode.}
  \item{anew}{
The new component of the first mode.}
}
\references{
Kroonenberg PM (1994) The TUCKALS line: a suite of programs for three-way data analysis. Computational Statistics and Data Analysis, 18, 73--96.
}
\author{
Rosaria Lombardo, Eric J Beh}
%\note{
%%  ~~further notes~~
%}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{internal}
%\keyword{component}% __ONLY ONE__ keyword per line
