\name{margK}
\alias{margK}
\title{
Tube marginals of a three-way contingency table
}
\description{
The function computes the tube marginals of the three-way contingency
table specified by the input parameter.
}
\usage{
margK(m)
}
\arguments{
  \item{m}{
The three-way contingency table.
}
}
%\details{
%%}
\value{
The tube marginals of the considered  three-way contingency table.
}

%\references{
%Beh EJ and Lombardo R (2014) Correspondence Analysis, Theory, Practice %and New Strategies. John Wiley & Sons.
%}
\author{
Rosaria Lombardo, Eric J Beh, Ida Camminatiello.
}

%\note{
%%  ~~further notes~~
%}

%% ~Make other sections like Warning with \section{Warning }{....} ~

%\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
%}
%\examples{
%##---- Should be DIRECTLY executable !! ----
%##-- ==>  Define data, use random,
%##--	or do  help(data=index)  for the standard data sets.
%}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{array}
\keyword{internal}% __ONLY ONE__ keyword per line
