\name{p.ext}
\alias{p.ext}
\title{
The external product in Tuckals3.
}
\description{The  computation of external product between the principal components.
}
\usage{
p.ext(x,y)
}
\arguments{
  \item{x}{
x matrix IxS}
\item{y}{
  y matrix JxS}
}
\value{
  resultant matrix (IxJ),S  with elements xis per yis
}
\references{
Beh EJ and Lombardo R (2014). Correspondence Analysis, Theory, Practice and New Strategies. John Wiley & Sons.
}
\author{
Rosaria Lombardo, Eric J Beh, Ida Camminatiello.
}
%\note{
%%  ~~further notes~~
%}
%\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
%}
%\examples{
%##---- Should be DIRECTLY executable !! ----
%##-- ==>  Define data, use random,
%##--	or do  help(data=index)  for the standard data sets.
%}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{tuckals3}
\keyword{product}% __ONLY ONE__ keyword per line
