\name{critera}
\alias{critera}
\title{
Criterion of the Tucker3 algorithm
}
\description{
The function critera is used to define the criterion of the Tucker3 algorithm.
}
\usage{
critera(aold, anew)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{aold}{
The old component of the first mode.}
  \item{anew}{
The new component of the first mode.}
}
\references{
Beh EJ and Lombardo R (2014). Correspondence Analysis, Theory, Practice and New Strategies. John Wiley & Sons.
}
\author{
Rosaria Lombardo, Eric J Beh}
%\note{
%%  ~~further notes~~
%}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{tucker}
\keyword{component}% __ONLY ONE__ keyword per line
